task/GraphQL-Apollo-Client : Fixed Code, Completed Task
