+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/Apollo/query/GetAllData.js
@@ -1,0 +1,12 @@
+import { gql } from "@apollo/client";
+
+export const GET_PASSENGERS = gql`
+  query MyQuery {
+    passengers {
+      id
+      name
+      age
+      gender
+    }
+  }
+`;
