+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/Apollo/query/GetDataById.js
@@ -1,16 +1,5 @@
 import { gql } from "@apollo/client";
 
-export const GET_PASSENGERS = gql`
-  query MyQuery {
-    passengers {
-      id
-      name
-      age
-      gender
-    }
-  }
-`;
-
 export const GET_PASSENGERS_BY_ID = gql`
   query MyQuery($id: Int!) {
     passengers(where: { id: { _eq: $id } }) {
