+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Home from "./Components/Home/Home";
+import Home from "./component/Home/Home";
 
 function App() {
   return (
