--- 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/Components/Passengers/ListPassenger.js
@@ -1,68 +1,0 @@
-import ListItem from "./ListItem";
-import { useLazyQuery, useQuery } from "@apollo/client";
-import LoadingAnimation from "../LoadingAnimation";
-import { GET_PASSENGERS_BY_ID } from "../../Apollo/constants";
-import { useState } from "react";
-
-const ListPassenger = (props) => {
-  const [getPassenger, { loading, error, data }] =
-    useLazyQuery(GET_PASSENGERS_BY_ID);
-  const [list, setList] = useState([]);
-  const [passengerId, setPassengerId] = useState(0);
-
-  if (loading) {
-    return <LoadingAnimation />;
-  }
-
-  if (error) {
-    console.log(error);
-    return null;
-  }
-
-  const onGetPassenger = () => {
-    getPassenger({
-      variables: {
-        id: passengerId,
-      },
-    });
-    setList(data?.passengers);
-  };
-
-  const onGetPassengerId = (e) => {
-    if (e.target) {
-      setPassengerId(e.target.value);
-    }
-  };
-
-  return (
-    <>
-      <div>
-        <input
-          style={{ marginBottom: "20px" }}
-          value={passengerId}
-          onChange={onGetPassengerId}
-        />
-        <button onClick={onGetPassenger}>Get Passengers</button>
-        <table cellPadding="5px" cellSpacing="0" style={{ margin: "auto" }}>
-          <thead>
-            <tr bgcolor="red">
-              <td>Nama</td>
-              <td>Umur</td>
-              <td>Jenis Kelamin</td>
-              <td bgcolor="white" className="removeBorder"></td>
-            </tr>
-            {data?.passengers.map((item) => (
-              <ListItem
-                key={item.id}
-                data={item}
-                hapusPengunjung={props.hapusPengunjung}
-              />
-            ))}
-          </thead>
-        </table>
-      </div>
-    </>
-  );
-};
-
-export default ListPassenger;
