+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/component/Home/Header.js
@@ -1,4 +1,4 @@
-import "../Assets/Css/Home.css";
+import "./Home.css";
 const Header = () => {
   return (
     <div className="header">
