+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/component/Home/Home.js
@@ -1,6 +1,6 @@
-import PassengerInput from "../Passengers/PassengerInput";
-import ListPassenger from "../Passengers/ListPassenger";
-import Header from "../Header";
+import PassengerInput from "../Passanger/PassengerInput";
+import ListPassenger from "../Passanger/ListPassenger";
+import Header from "./Header";
 
 const Home = () => {
   return (
