+++ 22_GraphQL-Apollo-Client/GrapQL_Query_task_starter-main/src/component/Passanger/ListItem.js
@@ -1,4 +1,4 @@
-import "../../Assets/Css/Home.css";
+import "../Home/Home.css";
 const ListItem = (props) => {
   const { id, name, age, gender } = props.data;
